refactor(index): introduce lance-index-core crate with abstract index trait layer
Extracts the abstract trait layer for scalar indices into a new
`lance-index-core` crate (~17 deps vs 75+ in lance-index), enabling
index plugin authors to implement ScalarIndex without depending on
the full lance-index implementation crate.

Moved to lance-index-core: Index, IndexType, IndexParams, MetricsCollector,
ScalarIndex, AnyQuery, IndexStore/Reader/Writer, RowIdRemapper, SearchResult,
CreatedIndex, UpdateCriteria, OldIndexDataFilter, ScalarIndexParams,
BuiltinIndexType, TrainingCriteria/TrainingOrdering, IndexFile.

lance-index and lance-table re-export all moved types for backward
compatibility. FragReuseIndex/MemWalIndex trait impls move to lance-table
where those types are defined.


refactor(index): break lance-table → lance-index-core dependency

Remove the unintended `lance-table → lance-index-core` dependency introduced
in the previous commit. Key changes:

- Restore `IndexFile` as a standalone struct in `lance-table` (duplicate of
  the one in `lance-index-core`), removing the re-export that forced the dep
- Add `index_files_to_table()` / `table_files_to_index()` free functions in
  `lance-index` to convert between the two distinct `IndexFile` types
- Introduce newtype wrappers `FragReuseIndexHandle` and `MemWalIndexHandle` in
  `lance-index` so that `Index` / `RowIdRemapper` (now in `lance-index-core`)
  can be implemented on the `lance-table` types without violating orphan rules
- Remove the `lance-index-core` trait impls that were added directly on
  `FragReuseIndex` / `MemWalIndex` in `lance-table`
- Update all call sites in `lance`, `lance-index`, and `lance-namespace-impls`
  to use the handle newtypes and conversion functions
