+++ rust/lance-index-core/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "lance-index-core"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+readme = "README.md"
+description = "Core traits and types for Lance index plugins"
+keywords.workspace = true
+categories.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+async-trait.workspace = true
+arrow-array.workspace = true
+arrow-schema.workspace = true
+arrow-select.workspace = true
+bytes.workspace = true
+datafusion-common.workspace = true
+datafusion-expr.workspace = true
+datafusion.workspace = true
+futures.workspace = true
+lance-core.workspace = true
+lance-io.workspace = true
+lance-select.workspace = true
+prost-types.workspace = true
+roaring.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+
+[lints]
+workspace = true
