+++ rust/lance-index/src/mem_wal.rs
@@ -5,19 +5,31 @@
 //!
 //! The data structures and table-format logic live in
 //! [`lance_table::system_index::mem_wal`]; this module re-exports them and
-//! implements the local [`Index`] trait for [`MemWalIndex`].
+//! provides a newtype wrapper that implements the [`Index`] trait.
 
 use std::any::Any;
 use std::sync::Arc;
 
 use async_trait::async_trait;
-use lance_core::Error;
+use lance_core::Result;
+use lance_core::deepsize::DeepSizeOf;
 use roaring::RoaringBitmap;
 use serde::Serialize;
 
 pub use lance_table::system_index::mem_wal::*;
 
 use crate::{Index, IndexType};
+
+/// Newtype wrapping [`MemWalIndex`] so that `lance-index` can implement
+/// the `Index` trait (orphan rules prevent implementing it directly in
+/// `lance-table`).
+pub struct MemWalIndexHandle(pub Arc<MemWalIndex>);
+
+impl DeepSizeOf for MemWalIndexHandle {
+    fn deep_size_of_children(&self, context: &mut lance_core::deepsize::Context) -> usize {
+        self.0.deep_size_of_children(context)
+    }
+}
 
 #[derive(Serialize)]
 struct MemWalStatistics {
@@ -29,7 +41,7 @@ struct MemWalStatistics {
 }
 
 #[async_trait]
-impl Index for MemWalIndex {
+impl Index for MemWalIndexHandle {
     fn as_any(&self) -> &dyn Any {
         self
     }
@@ -38,23 +50,23 @@ impl Index for MemWalIndex {
         self
     }
 
-    fn statistics(&self) -> lance_core::Result<serde_json::Value> {
+    fn statistics(&self) -> Result<serde_json::Value> {
         let stats = MemWalStatistics {
-            num_shards: self.details.num_shards,
-            num_merged_generations: self.details.merged_generations.len(),
-            num_shard_specs: self.details.sharding_specs.len(),
-            num_maintained_indexes: self.details.maintained_indexes.len(),
-            num_index_catchup_entries: self.details.index_catchup.len(),
+            num_shards: self.0.details.num_shards,
+            num_merged_generations: self.0.details.merged_generations.len(),
+            num_shard_specs: self.0.details.sharding_specs.len(),
+            num_maintained_indexes: self.0.details.maintained_indexes.len(),
+            num_index_catchup_entries: self.0.details.index_catchup.len(),
         };
         serde_json::to_value(stats).map_err(|e| {
-            Error::internal(format!(
+            lance_core::Error::internal(format!(
                 "failed to serialize MemWAL index statistics: {}",
                 e
             ))
         })
     }
 
-    async fn prewarm(&self) -> lance_core::Result<()> {
+    async fn prewarm(&self) -> Result<()> {
         Ok(())
     }
 
@@ -62,7 +74,7 @@ impl Index for MemWalIndex {
         IndexType::MemWal
     }
 
-    async fn calculate_included_frags(&self) -> lance_core::Result<RoaringBitmap> {
+    async fn calculate_included_frags(&self) -> Result<RoaringBitmap> {
         Ok(RoaringBitmap::new())
     }
 }
