+++ rust/lance-index/src/scalar/lance_format.rs
@@ -13,11 +13,8 @@ use lance_core::deepsize::DeepSizeOf;
 use lance_core::{Error, Result, cache::LanceCache};
 use lance_encoding::decoder::{DecoderPlugins, FilterExpression};
 use lance_encoding::version::LanceFileVersion;
-use lance_file::previous::{
-    reader::FileReader as PreviousFileReader,
-    writer::{FileWriter as PreviousFileWriter, ManifestProvider as PreviousManifestProvider},
-};
-use lance_file::reader::{self as current_reader, FileReaderOptions, ReaderProjection};
+use lance_file::previous::reader::FileReader as PreviousFileReader;
+use lance_file::reader::{FileReader as CurrentFileReader, FileReaderOptions, ReaderProjection};
 use lance_file::writer as current_writer;
 use lance_io::scheduler::{ScanScheduler, SchedulerConfig};
 use lance_io::utils::CachedFileSize;
@@ -130,34 +127,6 @@ impl LanceIndexStore {
     }
 }
 
-#[async_trait]
-impl<M: PreviousManifestProvider + Send + Sync> IndexWriter for PreviousFileWriter<M> {
-    async fn write_record_batch(&mut self, batch: RecordBatch) -> Result<u64> {
-        let offset = self.tell().await?;
-        self.write(&[batch]).await?;
-        Ok(offset as u64)
-    }
-
-    async fn finish(&mut self) -> Result<IndexFile> {
-        let summary = Self::finish(self).await?;
-        Ok(IndexFile {
-            path: String::new(),
-            size_bytes: summary.size_bytes,
-        })
-    }
-
-    async fn finish_with_metadata(
-        &mut self,
-        metadata: HashMap<String, String>,
-    ) -> Result<IndexFile> {
-        let summary = Self::finish_with_metadata(self, &metadata).await?;
-        Ok(IndexFile {
-            path: String::new(),
-            size_bytes: summary.size_bytes,
-        })
-    }
-}
-
 struct LanceIndexWriter {
     path: String,
     inner: current_writer::FileWriter,
@@ -198,10 +167,14 @@ impl IndexWriter for LanceIndexWriter {
     }
 }
 
+/// Newtype wrapper to allow implementing IndexReader for PreviousFileReader (a foreign type)
+struct PreviousIndexReader(PreviousFileReader);
+
 #[async_trait]
-impl IndexReader for PreviousFileReader {
+impl IndexReader for PreviousIndexReader {
     async fn read_record_batch(&self, offset: u64, _batch_size: u64) -> Result<RecordBatch> {
-        self.read_batch(offset as i32, ReadBatchParams::RangeFull, self.schema())
+        self.0
+            .read_batch(offset as i32, ReadBatchParams::RangeFull, self.0.schema())
             .await
     }
 
@@ -211,36 +184,39 @@ impl IndexReader for PreviousFileReader {
         projection: Option<&[&str]>,
     ) -> Result<RecordBatch> {
         let projection = match projection {
-            Some(projection) => self.schema().project(projection)?,
-            None => self.schema().clone(),
+            Some(projection) => self.0.schema().project(projection)?,
+            None => self.0.schema().clone(),
         };
-        self.read_range(range, &projection).await
+        self.0.read_range(range, &projection).await
     }
 
     async fn num_batches(&self, _batch_size: u64) -> u32 {
-        self.num_batches() as u32
+        self.0.num_batches() as u32
     }
 
     fn num_rows(&self) -> usize {
-        self.len()
+        self.0.len()
     }
 
     fn schema(&self) -> &lance_core::datatypes::Schema {
-        Self::schema(self)
+        PreviousFileReader::schema(&self.0)
     }
 }
 
+/// Newtype wrapper to allow implementing IndexReader for CurrentFileReader (a foreign type)
+struct CurrentIndexReader(CurrentFileReader);
+
 #[async_trait]
-impl IndexReader for current_reader::FileReader {
+impl IndexReader for CurrentIndexReader {
     async fn read_record_batch(&self, offset: u64, batch_size: u64) -> Result<RecordBatch> {
         let start = offset * batch_size;
         let end = start + batch_size;
-        let end = end.min(self.num_rows());
+        let end = end.min(self.0.num_rows());
         self.read_range(start as usize..end as usize, None).await
     }
 
     async fn read_global_buffer(&self, n: u32) -> Result<Bytes> {
-        Self::read_global_buffer(self, n).await
+        CurrentFileReader::read_global_buffer(&self.0, n).await
     }
 
     async fn read_range(
@@ -250,19 +226,20 @@ impl IndexReader for current_reader::FileReader {
     ) -> Result<RecordBatch> {
         if range.is_empty() {
             return Ok(RecordBatch::new_empty(Arc::new(
-                self.schema().as_ref().into(),
+                self.0.schema().as_ref().into(),
             )));
         }
         let projection = if let Some(projection) = projection {
             ReaderProjection::from_column_names(
-                self.metadata().version(),
-                self.schema(),
+                self.0.metadata().version(),
+                self.0.schema(),
                 projection,
             )?
         } else {
-            ReaderProjection::from_whole_schema(self.schema(), self.metadata().version())
+            ReaderProjection::from_whole_schema(self.0.schema(), self.0.metadata().version())
         };
         let batches = self
+            .0
             .read_stream_projected(
                 ReadBatchParams::Range(range),
                 u32::MAX,
@@ -284,7 +261,7 @@ impl IndexReader for current_reader::FileReader {
     ) -> Result<RecordBatch> {
         let empty_batch = || {
             Ok(RecordBatch::new_empty(Arc::new(
-                self.schema().as_ref().into(),
+                self.0.schema().as_ref().into(),
             )))
         };
         if ranges.is_empty() {
@@ -292,12 +269,12 @@ impl IndexReader for current_reader::FileReader {
         }
         let projection = if let Some(projection) = projection {
             ReaderProjection::from_column_names(
-                self.metadata().version(),
-                self.schema(),
+                self.0.metadata().version(),
+                self.0.schema(),
                 projection,
             )?
         } else {
-            ReaderProjection::from_whole_schema(self.schema(), self.metadata().version())
+            ReaderProjection::from_whole_schema(self.0.schema(), self.0.metadata().version())
         };
         // `DecodeBatchScheduler::schedule_ranges` requires sorted,
         // non-overlapping ranges; sort internally and permute the
@@ -311,6 +288,7 @@ impl IndexReader for current_reader::FileReader {
             .collect();
         let total_rows: u64 = sorted_ranges.iter().map(|r| r.end - r.start).sum();
         let batches = self
+            .0
             .read_stream_projected(
                 ReadBatchParams::Ranges(sorted_ranges),
                 (total_rows as u32).max(1),
@@ -363,47 +341,48 @@ impl IndexReader for current_reader::FileReader {
     ) -> Result<Pin<Box<dyn lance_io::stream::RecordBatchStream>>> {
         if range.is_empty() {
             return Ok(Box::pin(lance_io::stream::RecordBatchStreamAdapter::new(
-                Arc::new(self.schema().as_ref().into()),
+                Arc::new(self.0.schema().as_ref().into()),
                 futures::stream::empty(),
             )));
         }
         let projection = if let Some(projection) = projection {
             ReaderProjection::from_column_names(
-                self.metadata().version(),
-                self.schema(),
+                self.0.metadata().version(),
+                self.0.schema(),
                 projection,
             )?
         } else {
-            ReaderProjection::from_whole_schema(self.schema(), self.metadata().version())
+            ReaderProjection::from_whole_schema(self.0.schema(), self.0.metadata().version())
         };
-        self.read_stream_projected(
-            ReadBatchParams::Range(range),
-            4096,
-            2,
-            projection,
-            FilterExpression::no_filter(),
-        )
-        .await
+        self.0
+            .read_stream_projected(
+                ReadBatchParams::Range(range),
+                4096,
+                2,
+                projection,
+                FilterExpression::no_filter(),
+            )
+            .await
     }
 
     // V2 format has removed the row group concept,
     // so here we assume each batch is with 4096 rows.
     async fn num_batches(&self, batch_size: u64) -> u32 {
-        Self::num_rows(self).div_ceil(batch_size) as u32
+        CurrentFileReader::num_rows(&self.0).div_ceil(batch_size) as u32
     }
 
     fn num_rows(&self) -> usize {
-        Self::num_rows(self) as usize
+        CurrentFileReader::num_rows(&self.0) as usize
     }
 
     fn schema(&self) -> &lance_core::datatypes::Schema {
-        Self::schema(self)
+        CurrentFileReader::schema(&self.0)
     }
 
     fn file_size_bytes(&self) -> Option<u64> {
         // The manifest records each index file's size and passes it to the reader
         // at open, so it's already in metadata here (no extra I/O).
-        Some(self.metadata().file_size())
+        Some(self.0.metadata().file_size())
     }
 }
 
@@ -464,7 +443,7 @@ impl IndexStore for LanceIndexStore {
             .scheduler
             .open_file_with_priority(&path, self.io_priority, &cached_size)
             .await?;
-        match current_reader::FileReader::try_open(
+        match CurrentFileReader::try_open(
             file_scheduler,
             None,
             Arc::<DecoderPlugins>::default(),
@@ -473,7 +452,7 @@ impl IndexStore for LanceIndexStore {
         )
         .await
         {
-            Ok(reader) => Ok(Arc::new(reader)),
+            Ok(reader) => Ok(Arc::new(CurrentIndexReader(reader))),
             Err(e) => {
                 // If the error is a version conflict we can try to read the file with v1 reader
                 if let Error::VersionConflict { .. } = e {
@@ -484,7 +463,7 @@ impl IndexStore for LanceIndexStore {
                         Some(&self.metadata_cache),
                     )
                     .await?;
-                    Ok(Arc::new(file_reader))
+                    Ok(Arc::new(PreviousIndexReader(file_reader)))
                 } else {
                     Err(e)
                 }
@@ -558,7 +537,14 @@ impl IndexStore for LanceIndexStore {
     }
 
     async fn list_files_with_sizes(&self) -> Result<Vec<IndexFile>> {
-        list_index_files_with_sizes(&self.object_store, &self.index_dir).await
+        let files = list_index_files_with_sizes(&self.object_store, &self.index_dir).await?;
+        Ok(files
+            .into_iter()
+            .map(|f| IndexFile {
+                path: f.path,
+                size_bytes: f.size_bytes,
+            })
+            .collect())
     }
 }
